python_module_requirements.txt
scapy~=2.5.0
python-dotenv~=1.0.1
grpcio-tools~=1.62.0